get response but undefined when trying to json data
