+++ client/src/utils/API.js
@@ -1,6 +1,6 @@
-export const searchJobApi = (query) => {
+export const searchJobApi = (searchInput) => {
 
-    const url = `https://jsearch.p.rapidapi.com/search?query=${query}&page=1&num_pages=1&remote_jobs_only=true`;
+    const url = `https://jsearch.p.rapidapi.com/search?query=${searchInput}&page=1&num_pages=1&remote_jobs_only=true`;
     const options = {
 	method: 'GET',
 	headers: {
